refactor: rename utils/notifications to utils/backlog
- Rename utils/notifications.ts to utils/backlog.ts for better categorization
- Update import path in NotificationItem component
- Maintain all existing functionality and exports
- Better alignment with backlog-specific utilities
